+++ pkg/deployment/resources/pod_creator_gateway_container.go
@@ -131,18 +131,26 @@ func (a *ArangoGatewayContainer) GetEnvs() ([]core.EnvVar, []core.EnvFromSource)
 
 	envs.Add(true, k8sutil.GetLifecycleEnv()...)
 
-	var cmChecksum = ""
-
-	if cm, ok := a.cachedStatus.ConfigMap().V1().GetSimple(GetGatewayConfigMapName(a.input.ApiObject.GetName())); ok {
-		if v, ok := cm.Data[GatewayConfigChecksumFileName]; ok {
-			cmChecksum = v
+	if cm, ok := a.cachedStatus.ConfigMap().V1().GetSimple(GetGatewayConfigMapName(a.input.ArangoMember.GetName())); ok {
+		if v, ok := cm.Data[ConfigMapChecksumKey]; ok {
+			envs.Add(true, core.EnvVar{
+				Name:  MemberConfigChecksumENV,
+				Value: v,
+			})
 		}
 	}
 
-	envs.Add(true, core.EnvVar{
-		Name:  GatewayConfigChecksumENV,
-		Value: cmChecksum,
-	})
+	if !a.spec.Gateway.IsDynamic() {
+		if cm, ok := a.cachedStatus.ConfigMap().V1().GetSimple(GetGatewayConfigMapName(a.input.ApiObject.GetName())); ok {
+			if v, ok := cm.Data[ConfigMapChecksumKey]; ok {
+				envs.Add(true, core.EnvVar{
+					Name:  GatewayConfigChecksumENV,
+					Value: v,
+				})
+			}
+		}
+
+	}
 
 	if len(a.groupSpec.Envs) > 0 {
 		for _, env := range a.groupSpec.Envs {