+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - (Feature) (Gateway) ArangoDB JWT Auth Integration
 - (Feature) Scheduler Handler
 - (Feature) (Gateway) ArangoDB Auth Token
+- (Feature) (Gateway) Dynamic Configuration
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries